requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
pyodbc==5.0.1

# LLM Integration
google-genai==0.3.0

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.1
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
httpx==0.26.0
pytest-asyncio==0.21.1
